Remove ReaderBuilder and move its stuff to Reader
