+++ lib/rom/reader.rb
@@ -30,6 +30,49 @@ module ROM
     # @api private
     attr_reader :mapper
 
+    # Builds a reader instance for the provided relation
+    #
+    # @param [Symbol] name of the root relation
+    # @param [Relation] relation that the reader will use
+    # @param [MapperRegistry] registry of mappers
+    # @param [Array<Symbol>] a list of method names exposed by the relation
+    #
+    # @return [Reader]
+    #
+    # @api private
+    def self.build(name, relation, mappers, method_names = [])
+      klass = build_class(relation, method_names)
+      klass.new(name, relation, mappers)
+    end
+
+    # Build a reader subclass for the relation
+    #
+    # This method defines public methods on the class narrowing down data access
+    # only to the methods exposed by a given relation
+    #
+    # @param [Relation] relation that the reader will use
+    # @param [Array<Symbol>] a list of method names exposed by the relation
+    #
+    # @return [Class]
+    #
+    # @api private
+    def self.build_class(relation, method_names)
+      klass_name = "#{Reader.name}[#{Inflecto.camelize(relation.name)}]"
+
+      ClassBuilder.new(name: klass_name, parent: Reader).call do |klass|
+        method_names.each do |method_name|
+          klass.class_eval <<-RUBY, __FILE__, __LINE__ + 1
+            def #{method_name}(*args, &block)
+              new_relation = relation.send(#{method_name.inspect}, *args, &block)
+              self.class.new(
+                new_path(#{method_name.to_s.inspect}), new_relation, mappers
+              )
+            end
+          RUBY
+        end
+      end
+    end
+
     # @api private
     def initialize(path, relation, mappers, mapper = nil)
       @path = path.to_s
