+++ spec/unit/rom/reader_spec.rb
@@ -10,6 +10,55 @@ describe ROM::Reader do
   let(:mappers) { ROM::MapperRegistry.new(users: mapper) }
   let(:mapper) { double('mapper', header: []) }
 
+  describe '.build' do
+    subject(:reader) do
+      ROM::Reader.build(name, relation, mappers, [:all])
+    end
+
+    before do
+      relation.instance_exec do
+        def name
+          'users'
+        end
+
+        def all(*_args)
+          find_all
+        end
+      end
+    end
+
+    it 'sets reader class name' do
+      expect(reader.class.name).to eql("ROM::Reader[Users]")
+    end
+
+    it 'defines methods from relation' do
+      block = proc {}
+
+      expect(relation).to receive(:all)
+        .with(1, &block)
+        .and_return([joe])
+
+      expect(mapper).to receive(:process)
+        .with([joe])
+        .and_yield(joe)
+
+      result = reader.all(1, &block)
+
+      expect(result.path).to eql('users.all')
+      expect(result.to_a).to eql([joe])
+    end
+
+    it 'raises error when relation does not respond to the method' do
+      expect { reader.not_here }
+        .to raise_error(ROM::NoRelationError, /not_here/)
+    end
+
+    it 'raises error when relation does not respond to the method with args' do
+      expect { reader.find_by_id(1) }
+        .to raise_error(ROM::NoRelationError, /find_by_id/)
+    end
+  end
+
   describe '#initialize' do
     it 'raises error when mapper cannot be found' do
       expect { ROM::Reader.new(:not_here, relation, mappers) }
